refactor(ios): move DefaultMediaUploader to its own file
Moved verbatim out of MediaUploadServer.swift ahead of extracting the
upload route from the server, so that diff shows only the extraction.
